fix circ dep, add toasts, more improvements
