+++ src/app/src/contexts/ToastContextDef.ts
@@ -1,7 +1,7 @@
 import { createContext } from 'react';
 import type { AlertColor } from '@mui/material';
 
-interface ToastContextType {
+export interface ToastContextType {
   showToast: (message: string, severity?: AlertColor) => void;
 }
