+++ src/redteam/shared.ts
@@ -1,22 +1,84 @@
-import { type Severity, type Plugin, riskCategorySeverityMap } from './constants';
-import type { RedteamPluginObject } from './types';
+import chalk from 'chalk';
+import * as fs from 'fs';
+import yaml from 'js-yaml';
+import { doEval } from '../commands/eval';
+import logger, { setLogCallback } from '../logger';
+import type Eval from '../models/eval';
+import { isRunningUnderNpx } from '../util';
+import { loadDefaultConfig } from '../util/config/default';
+import { doGenerateRedteam } from './commands/generate';
+import type { RedteamRunOptions } from './types';
 
-export function getRiskCategorySeverityMap(
-  plugins?: RedteamPluginObject[],
-): Record<Plugin, Severity> {
-  const overrides =
-    plugins?.reduce(
-      (acc, plugin) => {
-        if (plugin.severity) {
-          acc[plugin.id as Plugin] = plugin.severity;
-        }
-        return acc;
-      },
-      {} as Record<Plugin, Severity>,
-    ) || {};
+export async function doRedteamRun(options: RedteamRunOptions): Promise<Eval | undefined> {
+  if (options.logCallback) {
+    setLogCallback(options.logCallback);
+  }
 
-  return {
-    ...riskCategorySeverityMap,
-    ...overrides,
-  };
+  let configPath: string | undefined = options.config || 'promptfooconfig.yaml';
+  let redteamPath = options.output || 'redteam.yaml';
+
+  if (options.liveRedteamConfig) {
+    // Write liveRedteamConfig to a temporary file
+    const tmpFile = fs.mkdtempSync('/tmp/redteam-') + '/redteam.yaml';
+    fs.writeFileSync(tmpFile, yaml.dump(options.liveRedteamConfig));
+    redteamPath = tmpFile;
+    // Do not use default config.
+    configPath = tmpFile;
+    logger.debug(`Using live config from ${tmpFile}`);
+    logger.debug(`Live config: ${JSON.stringify(options.liveRedteamConfig, null, 2)}`);
+  }
+
+  // Generate new test cases
+  logger.info('Generating test cases...');
+  await doGenerateRedteam({
+    ...options,
+    config: configPath,
+    output: redteamPath,
+    force: options.force,
+    inRedteamRun: true,
+    abortSignal: options.abortSignal,
+  });
+
+  // Check if redteam.yaml exists before running evaluation
+  if (!fs.existsSync(redteamPath)) {
+    logger.info('No test cases generated. Skipping scan.');
+    return;
+  }
+
+  // Run evaluation
+  logger.info('Running scan...');
+  const { defaultConfig } = await loadDefaultConfig();
+  const evalResult = await doEval(
+    {
+      ...options,
+      config: [redteamPath],
+      output: options.output ? [options.output] : undefined,
+      cache: true,
+      write: true,
+      filterProviders: options.filterProviders,
+      filterTargets: options.filterTargets,
+    },
+    defaultConfig,
+    redteamPath,
+    {
+      showProgressBar: true,
+      abortSignal: options.abortSignal,
+    },
+  );
+
+  logger.info(chalk.green('\nRed team scan complete!'));
+  const command = isRunningUnderNpx() ? 'npx promptfoo' : 'promptfoo';
+  if (options.liveRedteamConfig) {
+    logger.info(
+      `${chalk.blue('To view the results, click the "View Report" button or run: ')}${chalk.bold(`${command} redteam report`)} on the command line.`,
+    );
+  } else {
+    logger.info(
+      `${chalk.blue('To view the results, run: ')}${chalk.bold(`${command} redteam report`)}`,
+    );
+  }
+
+  // Clear the callback when done
+  setLogCallback(null);
+  return evalResult;
 }
