+++ src/redteam/shared_frontend.ts
@@ -1,0 +1,23 @@
+// This file is imported by the frontend and shouldn't use native dependencies.
+import { type Severity, type Plugin, riskCategorySeverityMap } from './constants';
+import type { RedteamPluginObject } from './types';
+
+export function getRiskCategorySeverityMap(
+  plugins?: RedteamPluginObject[],
+): Record<Plugin, Severity> {
+  const overrides =
+    plugins?.reduce(
+      (acc, plugin) => {
+        if (plugin.severity) {
+          acc[plugin.id as Plugin] = plugin.severity;
+        }
+        return acc;
+      },
+      {} as Record<Plugin, Severity>,
+    ) || {};
+
+  return {
+    ...riskCategorySeverityMap,
+    ...overrides,
+  };
+}
